+++ HISTORY.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+***Fixed:***
+
+- Patch Mypy when building with `pip` to fix access to build requirements
+
 ## 0.4 - 2022-01-04
 
 ***Added:***